Update double type to BigDecimal.

+++ src/test/java/com/imc/visitor/shapes/CircleTest.java
@@ -2,6 +2,8 @@ package com.imc.visitor.shapes;
 
 import org.junit.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -12,21 +14,21 @@ public class CircleTest {
     @Test
     public void shouldReturnACircleInstanceWhenRadiusIsValid() {
         // given
-        double radius = 5;
+        BigDecimal radius = BigDecimal.valueOf(5);
 
         // when
         Circle circle = Circle.builder().radius(radius).build();
 
         // then
         assertNotNull(circle);
-        assertTrue(circle.getRadius() == 5);
+        assertTrue(circle.getRadius().intValue() == 5);
 
     }
 
     @Test
     public void shouldThrowAnExceptionWhenRadiusIsNotPositive() {
         // given
-        double radius = -5;
+        BigDecimal radius = BigDecimal.valueOf(-5);
 
         // when
         Circle.CircleBuilder circleBuilder = Circle.builder();
